DD-1334 This commit adds a database schema and gets the module to a level of functionality. See commit desc.
This code is now functional but changes how the module works; it now maintains a database table which tracks when the links in each field for each node were checked.

By default, each node's fields will be checked no more-frequently than once per week.

We've also brought the default maximum run time down to a sane four minutes, so running a batch of link checks through the web UI should be more responsive.

Still to do is to use the Drupal batch UI to run an interactive link check job, to provide some feedback on how the process is going.

/Al

Signed-off-by: Alex Harries <alex@greyhead.co.uk>